Moving files and changing package
Just moved Lani's files to application folder.
