--- Magazine.java
@@ -1,17 +1,0 @@
-//Written by Aliilani McGuire
-
-package bookstore;
-
-public class Magazine {
-    public String title, pub_year;
-
-    public Magazine() {
-        title = "N/A";
-        pub_year = "1969";
-    }
-
-    public Magazine(String t, String py) {
-        title = t;
-        pub_year = py;
-    }
-}
