+++ application/Department.java
@@ -1,6 +1,6 @@
 //Written by Aliilani McGuire
 
-package bookstore;
+package application;
 
 public class Department {
     public String room_num, building, department_name;
