+++ application/JournalPaper.java
@@ -1,5 +1,16 @@
-package cs.CS3700.LMS.application;
+//Written by Aliilani McGuire
 
-public class JournalPaper {
-    
+package application;
+
+public class JournalPaper extends ResearchReport {
+    public String journal_title;
+
+    public JournalPaper() {
+        super();
+        journal_title = "N/A";
+    }
+    public JournalPaper(String t, String py, int an, String jt) {
+        super(t, py, jt);
+        journal_title = jt;
+    }
 }
