+++ application/Magazine.java
@@ -1,5 +1,17 @@
-package cs.CS3700.LMS.application;
+//Written by Aliilani McGuire
+
+package application;
 
 public class Magazine {
-    
+    public String title, pub_year;
+
+    public Magazine() {
+        title = "N/A";
+        pub_year = "1969";
+    }
+
+    public Magazine(String t, String py) {
+        title = t;
+        pub_year = py;
+    }
 }
