+++ application/Main.java
@@ -1,4 +1,4 @@
-package cs.CS3700.LMS.application;
+package application;
 // This Main Class is the foundation of the application. This is what will be used to create our GUI. Options for Java GUI seem to be:
 // Java Swing or JavaFX
